requirements.txt
websocket-client==1.8.0
requests==2.32.5
numpy==2.3.2
certifi==2025.8.3
